added exception handling

+++ src/main/java/com/example/springboot/service/UserSeviceImpl.java
@@ -1,16 +1,13 @@
 package com.example.springboot.service;
 
-import java.util.Arrays;
-import java.util.List;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.example.springboot.creator.UserCreator;
-import com.example.springboot.domain.UserRequest;
 import com.example.springboot.domain.UserResponse;
+import com.example.springboot.exeption.GenericResponse;
+import com.example.springboot.exeption.VacationAppException;
 import com.example.springboot.domain.User;
 import com.example.springboot.mapper.UserMapper;
 
@@ -24,23 +21,34 @@ public class UserSeviceImpl implements UserService {
 
 	@Override
 	public User getUser(String username) {
-		return userMapper.findUserByUniqueUsername(username);
+		logger.info("Zahtev za pregled usera = {} ", username);
+		User user = userMapper.findUserByUniqueUsername(username);
+		if (user != null) return userMapper.findUserByUniqueUsername(username);
+		else throw new VacationAppException("Korisnik sa username-om =" + username + " ne postoji u bazi", GenericResponse.USER_NOT_FOUND);
 	}	
 
 	@Override
 	public UserResponse updateUser(String username, String address, String name, String email) {
-		logger.debug("Update usera sa sl parametrima: username = {}, address = {}, name = {}, email = {} ", username, address, name, email);
-		userMapper.updateUser(username, address, name, email);	
-		//TODO: check how to solve this, when user does not exist
-		return new UserResponse(username, UserResponse.Status.SUCCESSFUL, "User uspesno updejtovan");
+		User user = userMapper.findUserByUniqueUsername(username);
+		if (user != null) {
+			userMapper.updateUser(username, address, name, email);
+			return new UserResponse(username, UserResponse.Status.SUCCESSFUL, "User uspesno updejtovan");			
+		}
+		else {
+			throw new VacationAppException("Korisnik sa username-om =" + username + " ne postoji u bazi", GenericResponse.USER_NOT_FOUND);
+		}
 	}	
 
 	@Override
 	public UserResponse deleteUser(String username) {
-		userMapper.deleteByUsername(username);
-		//TODO: check how to solve this, when user does not exist
-		return new UserResponse(username, UserResponse.Status.SUCCESSFUL, "User uspesno obrisan");
+		User user = userMapper.findUserByUniqueUsername(username);
+		if (user != null) {
+			userMapper.deleteByUsername(username);
+			return new UserResponse(username, UserResponse.Status.SUCCESSFUL, "User uspesno obrisan");			
+		}
+		else {
+			throw new VacationAppException("Korisnik sa username-om =" + username + " ne postoji u bazi", GenericResponse.USER_NOT_FOUND);
+		}		
 	}	
-
 
 }
